fix: resolve NestJS module export error
- Remove AuthController from exports in AuthModule (controllers cannot be exported)
- Move setCookies and parseTimeToMs methods from AuthController to AuthService
- Update NicknameGuard to use AuthService instead of AuthController
- Maintain proper separation of concerns: AuthService handles business logic, AuthController handles HTTP requests
- All tests passing (20/20)
